add some callable descriptions
implement get-vfunc-desc, list-vfuncs-desc for interfaces.
add get-callable-desc - use this for global functions

* src/function.lisp: (get-callable-desc) new method. for functions which
are in the "top level namespace" Eg.
(gir::get-callable-desc *gobject*  "signal_emitv")

* src/package.lisp: (get-callable-desc): export

* src/object.lisp: (get-vfunc-desc): (list-vfuncs-desc): implement for
intefaces. use interface-info-find-vfunc and interface-info-get-vfuncs.
NOTE,TODO This returns methods which are already found by
list-methods-desc and get-method-desc, and these are new methods are
added to see if the implementation of g-i makes a distinction between
the two and look for bugs in the g-i implementation