src/gvalue.lisp:  add comments
the definitions for g-value-get-interface and g-value-set-interface
were NOT added by mistake. These functions do get called at some point
when g-values are freed. TODO - document exactly when this happens

g-value-get-object and g-value-set-object autoconvrts to goject (and
now fake-gobject) - however there are use cases where we need to get
at the pointer to the gobject (which is what is being stored by these
calls in C). Add a comment to remind us about this.