Fix some shared-initialize definitions
SHARED-INITIALIZE gets called via UPDATE-INSTANCE-FOR-REDEFINED-TYPE
when the class gets redefined and MAKE-INSTANCES-OBSOLETE gets called
on extant objects.  In this case we should not reinitialize the object
since we are not creating it for the first time and don't have access
to the initargs which are supplied at creation time.

This bug is a shortcoming of the design which uses SHARED-INITIALIZE
for the purpose.

--
The error ccl was throwing is obscure.

If you recompile all files then objects such as
(setq *gobject* (require-namespace "GObject"))
become obsolete. ccl sets slots to nil.

Faslload shows this:

NIL has no slot named GIR::PTR.
   [Condition of type SIMPLE-ERROR]

  1: (SLOT-VALUE NIL GIR::PTR)
  2: (#<STANDARD-METHOD CFFI:TRANSLATE-TO-FOREIGN (T GIR::INFO-FFI)> NIL #<GIR::INFO-FFI #x3020017796CD>)
  3: (#:G-OBJECT-INFO-GET-PARENT NIL)
  4: (GIR:OBJECT-INFO-GET-PARENT NIL)
  5: (#<STANDARD-METHOD SHARED-INITIALIZE :AFTER (GIR:OBJECT-CLASS T)> #<error printing GIR:OBJECT-CLASS #x30200171571D> NIL :INFO NIL)
  9: (CCL::UPDATE-OBSOLETE-INSTANCE #<error printing GIR:OBJECT-CLASS #x30200171571D>)

* (require-namespace "GObject")
NIL is not a Lisp string or pointer.
   [Condition of type SIMPLE-ERROR]

Backtrace:
  0: (#<STANDARD-METHOD CFFI:TRANSLATE-TO-FOREIGN (T CFFI::FOREIGN-STRING-TYPE)> NIL #<FOREIGN-STRING-TYPE :UTF-8>)
  1: (G-IREPOSITORY-REQUIRE NIL NIL #<A Null Foreign Pointer> NIL #<A Foreign Pointer [stack-allocated] #x7EFD04742B70>)
  2: (REPOSITORY-REQUIRE NIL NIL #<A Null Foreign Pointer> NIL)
  3: (#<STANDARD-METHOD SHARED-INITIALIZE :AFTER (NAMESPACE T)> #N<GObject(2.0)> NIL :NAME NIL :VERSION NIL)
  4: (CCL::%%BEFORE-AND-AFTER-COMBINED-METHOD-DCODE (NIL #<STANDARD-METHOD SHARED-INITIALIZE :AFTER (NAMESPACE T)> . 17453100824688))
  5: (CCL::%%STANDARD-COMBINED-METHOD-DCODE (NIL (#<STANDARD-METHOD SHARED-INITIALIZE :AFTER (NAMESPACE T)>) #<CCL::STANDARD-KERNEL-METHOD SHARED-INITIALIZE (STANDARD-OBJECT T)>) 17453100824688)
  6: (NIL #<Unknown Arguments>)
  7: (CCL::UPDATE-OBSOLETE-INSTANCE #N<GObject(2.0)>)