src/gi-register-type.lisp: generate separate init-callbacks code
* src/gi-register-type.lisp:
(%gobject-subclassable-define-init-callbacks):
split into two functions for class and instance inits. The default
body is changed so each init function dispatches to a corresponding
user defined lisp function (if it exists)

This is not really a win:
Unfortunately because we are still in the constructor the lisp
function /has/ to operate on a cffi foreign pointer and not a
gobject. The lisp function will have to use cffi foreign funcall and
not gir:invoke.