generate code for vfuncs for user defined gobject classes
EXPERIMENTAL

* src/gi-register-type.lisp: (gobject-subclassable-info): extend with
a new field via :vfunc-overrides which is a list of cnames which will
be overridden by the user.

(%gobject-subclassable-get-name): now receives an optional FNAME
string fragment. for vfuncs this generates a name suitable for a
cffi:defcallback or a lisp function which the cffi:defcallback
dispatches to.  Construct the names in lisp style - after translating
underscores in cnames to hyphens.

(%gobject-subclassable-define-class-init-callback-form): generate code
to set the struct vfunc fields to the corresponding vfunc
cffi:defcallbacks. These would be defined via
%gobject-subclassable-define-vfunc-callbacks

(%gobject-subclassable-define-vfunc-callbacks): the boilerplate
code in the cffi:defcallback dispatches to a user defined function (if
defined) which is named as <DEFCALLBACK-NAME>-LISP

the extra lisp function is not really a win: Unfortunately because we
are still in the constructor the lisp function /has/ to operate on a
cffi foreign pointer and not a gobject. The lisp function will have to
call cffi:foreign funcall and not gir:invoke.