gstream.lisp: handle MKCL gray streams
NOTE trivial-gray-streams has to be updated for MKCL to use GRAY
as its implementation package.

"CLOSE" "OPEN-STREAM-P" "STREAM-ELEMENT-TYPE" are all functions in
MKCL and not generic methods. Shadowing-import these symbols from the
GRAY package into the TRIVIAL-GRAY-STREAM package for MKCL and then
shadow those symbols in the GIR-LIB package (where we define
overrides on the CL symbols in other lisps)