src/enum.lisp: compiler macro for (nget namespace ENUM :keyword)
to avoid multiple lookups.

TODO make sure it works
NOTE CCL fails to handle (funcall #'f args) as (f args)

;madhu 220214

Maybe Disable this for now?
This fails in ECL when trying to expand forms like

	    (if (featurep :wk)
		(gir:invoke (*gtk* "Window" "new")
			    (gir:nget *gtk* "WindowType" :toplevel))
		(gir:invoke (*gtk* "Window" "new")))

when :wk is not in the *features*

The correct thing to do with be load these components with these read
time conditionals  with :load-onlyt (or drastically to  use mk:oos
:load-source-instead-of-binary)

The compiler macro may also be the culprit in
"cl-gir.system: load enum.lisp early"